aria/apps/scheduler/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "aria-scheduler"
version = "1.0.0"
requires-python = ">=3.12"
dependencies = [
  "fastapi==0.115.2",
  "uvicorn[standard]==0.30.6",
  "pydantic==2.9.2",
  "temporalio==1.8.0",
  "redis==5.1.1",
  "httpx==0.27.2",
  "platform-adapters @ file://../../packages/platform-adapters",
  "opentelemetry-api==1.28.2",
  "opentelemetry-sdk==1.28.2",
  "opentelemetry-instrumentation-fastapi==0.49b2"
]
